Updated package structure for smallFullIXBRL mapping
