+++ document-generator-accounts/src/main/java/uk/gov/companieshouse/document/generator/accounts/mapper/smallfull/model/ixbrl/notes/Notes.java
@@ -1,0 +1,4 @@
+package uk.gov.companieshouse.document.generator.accounts.mapper.smallfull.model.ixbrl.notes;
+
+public class Notes {
+}
